Refactor ScriptChain and ScriptGroup
Both now rely on the RenderChainList control.
